+++ bundle/Entity/EzInfoCollection.php
@@ -7,7 +7,7 @@ use Doctrine\ORM\Mapping as ORM;
 /**
  * EzInfoCollection
  *
- * @ORM\Table(name="ezinfocollection")
+ * @ORM\Table(name="ezinfocollection", indexes={@ORM\Index(name="ezinfocollection_co_id_created", columns={"contentobject_id", "created"})})
  * @ORM\Entity(repositoryClass="Netgen\Bundle\InformationCollectionBundle\Repository\EzInfoCollectionRepository")
  */
 class EzInfoCollection