+++ doc/CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [master]
+### Added
+- `ezinfocollection` and `ezinfocollection_attribute` table indexes inside Entity classes
+
 ## [1.0] - 2017-17-02
 ### Added
 - ability to render recipient, sender and subject from template.