+++ doc/CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [master]
+## [1.0.1] - 2017-02-03
 ### Added
 - `ezinfocollection` and `ezinfocollection_attribute` table indexes inside Entity classes
 - bundle now requires `doctrine/orm` by default